Restructured Project. Stationary detection complete
