+++ .gitignore
@@ -41,4 +41,4 @@ jspm_packages
 .idea/*
 
 # Debug Output
-pedestrian_results
+images/jpg/pedestrian_results/*
