+++ js/intruder-detection.js
@@ -1,11 +1,3 @@
-/**
- * Created by cianduffy on 16/02/2017.
- */
-var express = require('express');
-var app = express();
-var server = require('http').createServer(app);
-server.listen(3000);
-var io = require('socket.io').listen(server);
 var opn = require('opn');
 var arDrone = require('ar-drone');
 var client  = arDrone.createClient();
@@ -19,14 +11,9 @@ var count = 0;
 var thickness = 2;
 var COLOUR_GREEN = [0, 255, 0];
 
-console.log('Waiting for PNGs');
 pngStream.on('error', console.log);
 pngStream.on('data', function (pngBuffer) {
-
-    console.log('PNG arrived');
-
     lastPNG = pngBuffer;
-
     cv.readImage(lastPNG, function (error, image) {
         if (error) throw error;
         if (image.width() < 1 || image.height() < 1) throw new Error('Image has no size');
@@ -39,13 +26,15 @@ pngStream.on('data', function (pngBuffer) {
             }
 
             if (people.length > 0) {
-                image.save('./pedestrian_results/fullbody-detection' + count + '.png');
-                console.log('Image saved to ./pedestrian_results/fullbody-detection' + count + '.png');
+                image.save('./images/jpg/pedestrian_results/fullbody-detection' + count + '.png');
+                console.log('Image saved to ./images/jpg/pedestrian_results/fullbody-detection' + count + '.png');
                 count ++;
                 if (!browserOpened) {
-                    opn('localhost:3000');
+                    opn('http://localhost:3000/intruder_detected');
+                    browserOpened = true;
                 }
             }
         });
     });
 });
+
