+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "ar-drone": "^0.3.3",
     "bower": "^1.8.0",
+    "child_process": "^1.0.2",
     "dronestream": "^1.1.1",
     "express": "^4.14.0",
     "nodecopter": "0.0.2",
